requirements.txt
# Créé par Viple SAS
Pillow>=8.0.0
pystray>=0.19.0